Minor big fix git add .

+++ InsuranceMicroservice/src/main/java/com/example/model/AllItemsInCartResponse.java
@@ -14,6 +14,6 @@ import lombok.ToString;
 @AllArgsConstructor
 @ToString
 public class AllItemsInCartResponse {
-    private Integer userId;
-    private List<Integer> listOfPolicyIds;
+    private Long userId;
+    private List<Long> listOfPolicyIds;
 }
